+++ swing/common-model/src/test/java/is/codion/swing/common/model/component/table/DefaultFilterTableSelectionModelTest.java
@@ -50,7 +50,7 @@ public class DefaultFilterTableSelectionModelTest {
 												return data.get(identifier);
 											}
 										})
-										.items(() -> data)
+										.supplier(() -> data)
 										.build();
 		tableModel.refresh();